* Fix error in `ernie-neo4j` fstab * Fix Docker hardening for machines with no Docker installed

+++ Technical/harden-and-update/checks-other/other.01-docker.sh
@@ -1,8 +1,11 @@
 #!/usr/bin/env bash
 set -e
 set -o pipefail
-echo "other.01 Docker: remove all unused containers, networks, images"
-echo -e "____CHECK____"
-docker system prune --all
-echo "Check PASSED"
-printf "\n\n"
+if command -v docker; then
+  echo "other.01 Docker: remove all unused containers, networks, images"
+  echo -e "____CHECK____"
+
+  docker system prune --all
+  echo "Check PASSED"
+  printf "\n\n"
+fi
